+++ src/utils/WinDynCalls.cpp
@@ -25,6 +25,7 @@ Sig_GetDpiForWindow DynGetDpiForWindow = nullptr;
 Sig_GetThreadDpiAwarenessContext DynGetThreadDpiAwarenessContext = nullptr;
 Sig_GetAwarenessFromDpiAwarenessContext DynGetAwarenessFromDpiAwarenessContext = nullptr;
 Sig_SetThreadDpiAwarenessContext DynSetThreadDpiAwarenessContext = nullptr;
+Sig_SystemParametersInfoForDpi DynSystemParametersInfoForDpi = nullptr;
 
 #define API_LOAD(name) Dyn##name = (Sig_##name)GetProcAddress(h, #name);
 
@@ -64,6 +65,8 @@ void InitDynCalls() {
         (Sig_GetAwarenessFromDpiAwarenessContext)GetProcAddress(h, "GetAwarenessFromDpiAwarenessContext");
     DynSetThreadDpiAwarenessContext =
         (Sig_SetThreadDpiAwarenessContext)GetProcAddress(h, "SetThreadDpiAwarenessContext");
+    DynSystemParametersInfoForDpi =
+        (Sig_SystemParametersInfoForDpi)GetProcAddress(h, "SystemParametersInfoForDpi");
 
     h = SafeLoadLibrary("uxtheme.dll");
     if (h) {